quick camera == camera view
