+++ res/layout/quick_attachment_drawer.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <merge xmlns:android="http://schemas.android.com/apk/res/android">
 
-    <org.thoughtcrime.securesms.components.camera.QuickCamera
+    <org.thoughtcrime.securesms.components.camera.CameraView
         android:id="@+id/quick_camera"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
