+++ src/org/thoughtcrime/securesms/components/camera/CameraView.java
@@ -20,11 +20,13 @@ import android.app.Activity;
 import android.content.Context;
 import android.content.pm.ActivityInfo;
 import android.graphics.Color;
+import android.graphics.Rect;
 import android.hardware.Camera;
 import android.hardware.Camera.CameraInfo;
 import android.hardware.Camera.Parameters;
 import android.hardware.Camera.PreviewCallback;
 import android.hardware.Camera.Size;
+import android.os.AsyncTask;
 import android.os.Build;
 import android.os.Build.VERSION;
 import android.support.annotation.NonNull;
@@ -38,12 +40,13 @@ import android.view.Surface;
 import android.widget.FrameLayout;
 
 import java.io.IOException;
+import java.util.List;
 
-import com.commonsware.cwac.camera.CameraHost;
 import com.commonsware.cwac.camera.CameraHost.FailureReason;
 import com.commonsware.cwac.camera.CameraUtils;
 
 import org.thoughtcrime.securesms.ApplicationContext;
+import org.thoughtcrime.securesms.util.BitmapUtil;
 import org.thoughtcrime.securesms.util.Util;
 import org.whispersystems.jobqueue.Job;
 import org.whispersystems.jobqueue.JobParameters;
@@ -53,15 +56,16 @@ import org.whispersystems.libaxolotl.util.guava.Optional;
 public class CameraView extends FrameLayout {
   private static final String TAG = CameraView.class.getSimpleName();
 
+  private final CameraSurfaceView   surface;
   private final OnOrientationChange onOrientationChange;
 
-  private @NonNull volatile Optional<Camera> camera = Optional.absent();
+  private @NonNull volatile Optional<Camera> camera   = Optional.absent();
+  private          volatile int              cameraId = CameraInfo.CAMERA_FACING_BACK;
 
-  private CameraHost        host;
-  private CameraSurfaceView surface;
-
-  private int displayOrientation     = -1;
-  private int outputOrientation      = -1;
+  private           boolean            started;
+  private @Nullable CameraViewListener listener;
+  private           int                displayOrientation     = -1;
+  private           int                outputOrientation      = -1;
 
   public CameraView(Context context) {
     this(context, null);
@@ -75,23 +79,23 @@ public class CameraView extends FrameLayout {
     super(context, attrs, defStyle);
     setBackgroundColor(Color.BLACK);
 
+    if (isMultiCamera()) cameraId = CameraInfo.CAMERA_FACING_FRONT;
+
     surface             = new CameraSurfaceView(getContext());
     onOrientationChange = new OnOrientationChange(context.getApplicationContext());
     addView(surface);
   }
 
-  public void setHost(@NonNull CameraHost host) {
-    this.host = host;
-  }
-
   @TargetApi(Build.VERSION_CODES.ICE_CREAM_SANDWICH)
   public void onResume() {
+    if (started) return;
+    started = true;
     Log.w(TAG, "onResume() queued");
     enqueueTask(new SerialAsyncTask<Pair<Camera, FailureReason>>() {
       @Override protected Pair<Camera, FailureReason> onRunBackground() {
         try {
-          if (host.getCameraId() >= 0) {
-            return new Pair<>(Camera.open(host.getCameraId()), null);
+          if (cameraId >= 0) {
+            return new Pair<>(Camera.open(cameraId), null);
           } else {
             return new Pair<>(null, FailureReason.NO_CAMERAS_REPORTED);
           }
@@ -102,7 +106,7 @@ public class CameraView extends FrameLayout {
 
       @Override protected void onPostMain(Pair<Camera, FailureReason> result) {
         if (result.first == null) {
-          host.onCameraFail(result.second);
+          if (listener != null) listener.onCameraFail(result.second);
           return;
         }
 
@@ -128,6 +132,8 @@ public class CameraView extends FrameLayout {
   }
 
   public void onPause() {
+    if (!started) return;
+    started = false;
     Log.w(TAG, "onPause() queued");
     final Optional<Camera> cameraToDestroy = camera;
 
@@ -151,6 +157,10 @@ public class CameraView extends FrameLayout {
         Log.w(TAG, "onPause() completed");
       }
     });
+  }
+
+  public boolean isStarted() {
+    return started;
   }
 
   @Override
@@ -181,7 +191,7 @@ public class CameraView extends FrameLayout {
 
     if (camera.isPresent()) {
       final Size previewSize = camera.get().getParameters().getPreviewSize();
-      if (getDisplayOrientation() == 90 || getDisplayOrientation() == 270) {
+      if (displayOrientation == 90 || displayOrientation == 270) {
         previewWidth  = previewSize.height;
         previewHeight = previewSize.width;
       } else {
@@ -211,19 +221,51 @@ public class CameraView extends FrameLayout {
     }
   }
 
-  public int getDisplayOrientation() {
-    return displayOrientation;
+  public void setListener(@Nullable CameraViewListener listener) {
+    this.listener = listener;
+  }
+
+  public boolean isMultiCamera() {
+    return Camera.getNumberOfCameras() > 1;
+  }
+
+  public boolean isRearCamera() {
+    return cameraId == CameraInfo.CAMERA_FACING_BACK;
+  }
+
+  public void flipCamera() {
+    enqueueTask(new PostInitializationTask() {
+      @Override protected void onPostMain(Object o) {
+        if (Camera.getNumberOfCameras() > 1) {
+          cameraId = cameraId == CameraInfo.CAMERA_FACING_FRONT
+                   ? CameraInfo.CAMERA_FACING_BACK
+                   : CameraInfo.CAMERA_FACING_FRONT;
+          onPause();
+          onResume();
+        }
+      }
+    });
   }
 
   public void setOneShotPreviewCallback(PreviewCallback callback) {
     if (camera.isPresent()) camera.get().setOneShotPreviewCallback(callback);
   }
 
+  @TargetApi(14)
   private void onCameraReady() {
     if (!camera.isPresent()) return;
 
-    final Parameters parameters = camera.get().getParameters();
+    final Parameters   parameters = camera.get().getParameters();
+    final List<String> focusModes = parameters.getSupportedFocusModes();
+
     if (VERSION.SDK_INT >= 14) parameters.setRecordingHint(true);
+
+    if (VERSION.SDK_INT >= 14 && focusModes.contains(Parameters.FOCUS_MODE_CONTINUOUS_PICTURE)) {
+      parameters.setFocusMode(Parameters.FOCUS_MODE_CONTINUOUS_PICTURE);
+    } else if (focusModes.contains(Parameters.FOCUS_MODE_CONTINUOUS_VIDEO)) {
+      parameters.setFocusMode(Parameters.FOCUS_MODE_CONTINUOUS_VIDEO);
+    }
+
     camera.get().setParameters(parameters);
 
     enqueueTask(new PostInitializationTask<Void>() {
@@ -248,7 +290,7 @@ public class CameraView extends FrameLayout {
 
     Size preferredSize = VERSION.SDK_INT > 11 ? camera.getParameters().getPreferredPreviewSizeForVideo() : null;
     if (preferredSize == null) {
-      preferredSize = CameraUtils.getBestAspectPreviewSize(getDisplayOrientation(),
+      preferredSize = CameraUtils.getBestAspectPreviewSize(displayOrientation,
                                                            getMeasuredWidth(),
                                                            getMeasuredHeight(),
                                                            parameters);
@@ -260,13 +302,11 @@ public class CameraView extends FrameLayout {
   private void startPreview() {
     if (camera.isPresent()) {
       camera.get().startPreview();
-      host.autoFocusAvailable();
     }
   }
 
   private void stopPreview() {
     if (camera.isPresent()) {
-      host.autoFocusUnavailable();
       camera.get().stopPreview();
     }
   }
@@ -318,7 +358,7 @@ public class CameraView extends FrameLayout {
 
   private @NonNull CameraInfo getCameraInfo() {
     final CameraInfo info = new Camera.CameraInfo();
-    Camera.getCameraInfo(host.getCameraId(), info);
+    Camera.getCameraInfo(cameraId, info);
     return info;
   }
 
@@ -373,6 +413,59 @@ public class CameraView extends FrameLayout {
         }
       }
     }
+  }
+
+  public void takePicture(final Rect previewRect) {
+    // TODO check camera state
+    if (!getCamera().isPresent() || getCamera().get().getParameters() == null) {
+      Log.w(TAG, "camera not in capture-ready state");
+      return;
+    }
+
+    setOneShotPreviewCallback(new Camera.PreviewCallback() {
+      @Override
+      public void onPreviewFrame(byte[] data, final Camera camera) {
+        final int  rotation     = getCameraPictureOrientation();
+        final Size previewSize  = camera.getParameters().getPreviewSize();
+        final Rect croppingRect = getCroppedRect(previewSize, previewRect, rotation);
+
+        Log.w(TAG, "previewSize: " + previewSize.width + "x" + previewSize.height);
+        Log.w(TAG, "data bytes: " + data.length);
+        Log.w(TAG, "previewFormat: " + camera.getParameters().getPreviewFormat());
+        Log.w(TAG, "croppingRect: " + croppingRect.toString());
+        Log.w(TAG, "rotation: " + rotation);
+        new RotatePreviewAsyncTask(previewSize, rotation, croppingRect).execute(data);
+      }
+    });
+  }
+
+  private Rect getCroppedRect(Size cameraPreviewSize, Rect visibleRect, int rotation) {
+    final int previewWidth  = cameraPreviewSize.width;
+    final int previewHeight = cameraPreviewSize.height;
+
+    if (rotation % 180 > 0) rotateRect(visibleRect);
+
+    float scale = (float) previewWidth / visibleRect.width();
+    if (visibleRect.height() * scale > previewHeight) {
+      scale = (float) previewHeight / visibleRect.height();
+    }
+    final float newWidth  = visibleRect.width()  * scale;
+    final float newHeight = visibleRect.height() * scale;
+    final float centerX   = (VERSION.SDK_INT < 14) ? previewWidth - newWidth / 2 : previewWidth / 2;
+    final float centerY   = previewHeight        / 2;
+
+    visibleRect.set((int) (centerX - newWidth  / 2),
+                    (int) (centerY - newHeight / 2),
+                    (int) (centerX + newWidth  / 2),
+                    (int) (centerY + newHeight / 2));
+
+    if (rotation % 180 > 0) rotateRect(visibleRect);
+    return visibleRect;
+  }
+
+  @SuppressWarnings("SuspiciousNameCombination")
+  private void rotateRect(Rect rect) {
+    rect.set(rect.top, rect.left, rect.bottom, rect.right);
   }
 
   private void enqueueTask(SerialAsyncTask job) {
@@ -436,5 +529,43 @@ public class CameraView extends FrameLayout {
     }
   }
 
+  private class RotatePreviewAsyncTask extends AsyncTask<byte[], Void, byte[]> {
+    private final Size previewSize;
+    private final int  rotation;
+    private final Rect croppingRect;
+
+    public RotatePreviewAsyncTask(Size previewSize, int rotation, Rect croppingRect) {
+      this.previewSize = previewSize;
+      this.rotation = rotation;
+      this.croppingRect = croppingRect;
+    }
+
+    @Override
+    protected byte[] doInBackground(byte[]... params) {
+      byte[] data = params[0];
+      try {
+
+        return BitmapUtil.createFromNV21(data,
+                                         previewSize.width,
+                                         previewSize.height,
+                                         rotation,
+                                         croppingRect);
+      } catch (IOException e) {
+        return null;
+      }
+    }
+
+    @Override
+    protected void onPostExecute(byte[] imageBytes) {
+      if (imageBytes != null && listener != null) listener.onImageCapture(imageBytes);
+    }
+  }
+
   private static class PreconditionsNotMetException extends Exception {}
+
+  public interface CameraViewListener {
+    void onImageCapture(@NonNull final byte[] imageBytes);
+
+    void onCameraFail(FailureReason reason);
+  }
 }
